Coverity Scan and Krazy fixes
Fixed problems reported by Coverity Scan
Fixed problems or quiet false positives reported by Krazy

Reviewed-on: https://codereview.kdab.com/12857

+++ src/view/quick/quickmaskedmousearea.cpp
@@ -116,7 +116,7 @@ void QuickMaskedMouseArea::setTolerance(qreal tolerance)
         return;
     }
 
-    m_tolerance = qMax(tolerance, 0.);
+    m_tolerance = qMax(tolerance, qreal(0.0));
     emit toleranceChanged(m_tolerance);
 }
